decaf377_frost
Module keys

FROST key shares and key generation.

Modules

dkg

Distributed key generation without a trusted dealer.

Functions

generate_with_dealer

Allows all participants’ keys to be generated using a central, trusted dealer.

reconstruct

Recompute the secret from t-of-n secret shares using Lagrange interpolation.

split

Splits an existing key into FROST shares.

Type Aliases

IdentifierList

The identifier list to use when generating key shares.

KeyPackage

A FROST keypair, which can be generated either by a trusted dealer or using a DKG.

PublicKeyPackage

Public data that contains all the signers’ public keys as well as the group public key.

SecretShare

Secret and public key material generated by a dealer performing generate_with_dealer.

SigningShare

A secret scalar value representing a signer’s share of the group secret.

VerifiableSecretSharingCommitment

Contains the commitments to the coefficients for our secret polynomial f, used to generate participants’ key shares.

VerifyingKey

A valid verifying key for Schnorr signatures over a FROST [Ciphersuite::Group].

VerifyingShare

A public group element that represents a single signer’s public verification share.