decaf377_frost::keys
Module dkg

Distributed key generation without a trusted dealer.

Modules

round1

DKG Round 1 structures.

round2

DKG Round 2 structures.

Functions

part1

Performs the first part of the distributed key generation protocol for the given participant.

part2

Performs the second part of the distributed key generation protocol for the participant holding the given round1::SecretPackage, given the received round1::Packages received from the other participants.

part3

Performs the third and final part of the distributed key generation protocol for the participant holding the given round2::SecretPackage, given the received round1::Packages and round2::Packages received from the other participants.