decaf377_frost::keys::dkg
Function part3

pub fn part3(
    round2_secret_package: &SecretPackage,
    round1_packages: &HashMap<Identifier, Package>,
    round2_packages: &HashMap<Identifier, Package>,
) -> Result<(KeyPackage, PublicKeyPackage), Error>

Performs the third and final part of the distributed key generation protocol for the participant holding the given round2::SecretPackage, given the received round1::Packages and round2::Packages received from the other participants.

It returns the KeyPackage that has the long-lived key share for the participant, and the PublicKeyPackages that has public information about all participants; both of which are required to compute FROST signatures.