decaf377_frost::keys
Type Alias VerifiableSecretSharingCommitment

pub type VerifiableSecretSharingCommitment = VerifiableSecretSharingCommitment<Decaf377Rdsa>;

Contains the commitments to the coefficients for our secret polynomial f, used to generate participants’ key shares.

VerifiableSecretSharingCommitment contains a set of commitments to the coefficients (which themselves are scalars) for a secret polynomial f, where f is used to generate each ith participant’s key share f(i). Participants use this set of commitments to perform verifiable secret sharing.

Note that participants MUST be assured that they have the same VerifiableSecretSharingCommitment, either by performing pairwise comparison, or by using some agreed-upon public location for publication, where each participant can ensure that they received the correct (and same) value.

Aliased Type

struct VerifiableSecretSharingCommitment(/* private fields */);