decaf377_frost
Module round2

Signing Round 2 functionality and types.

Structs

SignatureShare

A FROST participant’s signature share, which the Coordinator will aggregate with all other signer’s shares into the joint signature.

Functions

sign

Performed once by each participant selected for the signing operation.

sign_randomized

Like sign, but for producing signatures with a randomized verification key.