decaf377_frost
Type Alias Error

pub type Error = Error<Decaf377Rdsa>;

A FROST-related error.

Aliased Type

enum Error {

    InvalidMinSigners,
    InvalidMaxSigners,
    InvalidCoefficients,
    MalformedIdentifier,
    DuplicatedIdentifier,
    UnknownIdentifier,
    IncorrectNumberOfIdentifiers,
    MalformedSigningKey,
    MalformedVerifyingKey,
    MalformedSignature,
    InvalidSignature,
    DuplicatedShares,
    IncorrectNumberOfShares,
    IdentityCommitment,
    MissingCommitment,
    IncorrectCommitment,
    IncorrectNumberOfCommitments,
    InvalidSignatureShare {
        culprit: Identifier<Decaf377Rdsa>,
    },
    InvalidSecretShare,
    PackageNotFound,
    IncorrectNumberOfPackages,
    IncorrectPackage,
    DKGNotSupported,
    InvalidProofOfKnowledge {
        culprit: Identifier<Decaf377Rdsa>,
    },
    FieldError(FieldError),
    GroupError(GroupError),
    InvalidCoefficient,
    IdentifierDerivationNotSupported,
    SerializationError,
    DeserializationError,

}

Variants

InvalidMinSigners

min_signers is invalid

InvalidMaxSigners

max_signers is invalid

InvalidCoefficients

max_signers is invalid

MalformedIdentifier

This identifier is unserializable.

DuplicatedIdentifier

This identifier is duplicated.

UnknownIdentifier

This identifier does not belong to a participant in the signing process.

IncorrectNumberOfIdentifiers

Incorrect number of identifiers.

MalformedSigningKey

The encoding of a signing key was malformed.

MalformedVerifyingKey

The encoding of a verifying key was malformed.

MalformedSignature

The encoding of a signature was malformed.

InvalidSignature

Signature verification failed.

DuplicatedShares

Duplicated shares provided

IncorrectNumberOfShares

Incorrect number of shares.

IdentityCommitment

Commitment equals the identity

MissingCommitment

The participant’s commitment is missing from the Signing Package

IncorrectCommitment

The participant’s commitment is incorrect

IncorrectNumberOfCommitments

Incorrect number of commitments.

InvalidSignatureShare

Signature share verification failed.

Fields

culprit: Identifier<Decaf377Rdsa>

The identifier of the signer whose share validation failed.

InvalidSecretShare

Secret share verification failed.

PackageNotFound

Round 1 package not found for Round 2 participant.

IncorrectNumberOfPackages

Incorrect number of packages.

IncorrectPackage

The incorrect package was specified.

DKGNotSupported

The ciphersuite does not support DKG.

InvalidProofOfKnowledge

The proof of knowledge is not valid.

Fields

culprit: Identifier<Decaf377Rdsa>

The identifier of the signer whose share validation failed.

FieldError(FieldError)

Error in scalar Field.

GroupError(GroupError)

Error in elliptic curve Group.

InvalidCoefficient

Error in coefficient commitment deserialization.

IdentifierDerivationNotSupported

The ciphersuite does not support deriving identifiers from strings.

SerializationError

Error serializing value.

DeserializationError

Error deserializing value.