decaf377_frost::keys
Function split

pub fn split<R: RngCore + CryptoRng>(
    secret: &SigningKey<SpendAuth>,
    max_signers: u16,
    min_signers: u16,
    identifiers: IdentifierList<'_>,
    rng: &mut R,
) -> Result<(HashMap<Identifier, SecretShare>, PublicKeyPackage), Error>

Splits an existing key into FROST shares.

This is identical to generate_with_dealer but receives an existing key instead of generating a fresh one. This is useful in scenarios where the key needs to be generated externally or must be derived from e.g. a seed phrase.